requirements.txt
flask>=2.3.0
flask-login>=0.6.0
werkzeug>=2.3.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
